Add VehicleFactory

+++ Ex03.GarageLogic/ValueOutOfRangeExpection.cs
@@ -1,8 +1,4 @@
 ﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
 
 namespace Ex03.GarageLogic
 {
@@ -14,7 +10,7 @@ namespace Ex03.GarageLogic
         private float m_MinValue;
         public  float MinValue { get => m_MinValue; }
 
-        public ValueOutOfRangeExpection(int i_MinValue, int i_MaxValue) : base(String.Format("Value is out of range , range is between {0} to {1}",i_MinValue,i_MaxValue))
+        public ValueOutOfRangeExpection(float i_MinValue ,float i_MaxValue) : base(String.Format("Value is out of range , range is between {0} to {1}",i_MinValue,i_MaxValue))
         {
             m_MinValue = i_MinValue;
             m_MaxValue = i_MaxValue;
